requirements.txt
fastapi
uvicorn
requests
python-dotenv
apscheduler
textblob
